add and fix domain model
