+++ domain/src/main/java/com/rmakiyama/spatz/domain/model/ImageUrl.kt
@@ -1,3 +1,3 @@
 package com.rmakiyama.spatz.domain.model
 
-data class TweetId(val value: Long)
+data class ImageUrl(val value: String)
