--- domain/src/main/java/com/rmakiyama/spatz/domain/model/Tweet.kt
@@ -1,7 +1,0 @@
-package com.rmakiyama.spatz.domain.model
-
-data class Tweet(
-    val id: TweetId,
-    val text: TweetText,
-    val createdTimestamp: Long
-)
