--- domain/src/main/java/com/rmakiyama/spatz/domain/model/TweetText.kt
@@ -1,3 +1,0 @@
-package com.rmakiyama.spatz.domain.model
-
-data class TweetText(val value: String)
