+++ domain/src/main/java/com/rmakiyama/spatz/domain/model/tweet/Tweet.kt
@@ -1,0 +1,10 @@
+package com.rmakiyama.spatz.domain.model.tweet
+
+import com.rmakiyama.spatz.domain.model.user.User
+
+data class Tweet(
+    val id: TweetId,
+    val text: TweetText,
+    val user: User,
+    val createdTimestamp: Long
+)
