+++ domain/src/main/java/com/rmakiyama/spatz/domain/model/tweet/TweetId.kt
@@ -1,0 +1,3 @@
+package com.rmakiyama.spatz.domain.model.tweet
+
+data class TweetId(val value: Long)
