+++ domain/src/main/java/com/rmakiyama/spatz/domain/model/user/UserDescription.kt
@@ -1,0 +1,3 @@
+package com.rmakiyama.spatz.domain.model.user
+
+data class UserDescription(val value: String)
