+++ domain/src/main/java/com/rmakiyama/spatz/domain/model/user/UserScreenName.kt
@@ -1,0 +1,3 @@
+package com.rmakiyama.spatz.domain.model.user
+
+data class UserScreenName(val value: String)
