+++ domain/src/main/java/com/rmakiyama/spatz/domain/repository/TweetRepository.kt
@@ -1,6 +1,6 @@
 package com.rmakiyama.spatz.domain.repository
 
-import com.rmakiyama.spatz.domain.model.Tweet
+import com.rmakiyama.spatz.domain.model.tweet.Tweet
 
 interface TweetRepository {
