+++ domain/src/main/java/com/rmakiyama/spatz/domain/usecase/GetTweetsUseCase.kt
@@ -1,5 +1,5 @@
 package com.rmakiyama.spatz.domain.usecase
 
-import com.rmakiyama.spatz.domain.model.Tweet
+import com.rmakiyama.spatz.domain.model.tweet.Tweet
 
 interface GetTweetsUseCase : UseCase<Unit, List<Tweet>>
