+++ features/home/src/main/java/com/rmakiyama/spatz/home/HomeFragment.kt
@@ -5,7 +5,7 @@ import android.view.View
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.observe
-import com.rmakiyama.spatz.domain.model.Tweet
+import com.rmakiyama.spatz.domain.model.tweet.Tweet
 import dagger.hilt.android.AndroidEntryPoint
 import timber.log.Timber
